vmdisplay: create eglImage without the format modifier attributes
On some platform, failed to create NV12 format eglImage from dma-buf
with the format modifier attributes.

Per Dongwon's suggestion, the failure can be avoid by removing the
limiations. But the image's color is red if use `DRM_FORMAT_NV12`. The
color is correct if use two textures to represent Y and UV seperately.

Tracked-On: https://github.com/intel/ias/issues/62
Signed-off-by: Liu Xinyun <xinyun.liu@intel.com>